provision: Improved efficiency and insights
Calculating the results for a large number of votes takes time. Every time the results are viewed, this computation takes place, slowing down the response time. Instead, store the results separately in a database and call them whenever required.
This commit also creates a provision for voter turnout by gender as an insight.